requirements.txt
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pyaml==20.4.0
pycparser==2.20
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
